Изменены типы данных в сущностях на 4 байтовые варианты, для более корректного хранения в БД

+++ src/main/java/com/example/Pitlance/Models/SellerModelAndDTO/SellerNameEmailBalancePhonePasswordTPI.java
@@ -1,6 +1,8 @@
 package com.example.Pitlance.Models.SellerModelAndDTO;
 
-public record SellerNameEmailBalancePhonePasswordTPI(String companyName, String sellerName, String sellerLastName, String email, Integer phoneNumber, String taxPayerId, Integer balance, String password) {
+import java.math.BigDecimal;
+
+public record SellerNameEmailBalancePhonePasswordTPI(String companyName, String sellerName, String sellerLastName, String email, int phoneNumber, String taxPayerId, BigDecimal balance, String password) {
     public static SellerNameEmailBalancePhonePasswordTPI of(Seller seller){
         return new SellerNameEmailBalancePhonePasswordTPI(seller.getCompanyName(), seller.getSellerName(), seller.getSellerLastName(), seller.getEmail(), seller.getPhoneNumber(), seller.getTaxPayerId(), seller.getBalance(), seller.getPassword());
     }
